+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v.2.0.0
+
+- Renamed to sptables (sptables - Pure Iptables firewall for servers)
+- Still Docker compatible
+- Added standalone mode (can run both with or without Docker)
+- Slightly changed file structure
+
 ## v1.1.0
 
 - Added proxylist to bypass connection limits for certain ports. Intended to be used as trusted reverse proxy / CDN IP addresses list
@@ -8,4 +15,4 @@
 
 ## v1.0.0
 
-- Initial release
+- Initial release (Formerly: dtables - Docker compatible Iptables firewall)